pyproject.toml
[tool.poetry]
name = "testforproninteam"
version = "0.1.0"
description = ""
authors = ["Spaceocean <spaceocean@outlook.com>"]
readme = "README.md"
packages = [{include = "testforproninteam"}]


[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
djangorestframework = "^3.16.0"
drf-yasg = "^1.21.10"
pillow = "^11.2.1"
idna = "^3.10"
python-dotenv = "^1.1.0"
faker = "^37.1.0"
django-redis = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
init-env = "scripts.setup_env:main"